+++ src/Pagination/OffsetPagination.php
@@ -13,6 +13,7 @@ use Yiisoft\Data\Paginator\PaginatorInterface;
 use Yiisoft\Html\Html;
 use Yiisoft\Widget\Widget;
 
+use function array_key_exists;
 use function max;
 use function min;
 
@@ -365,9 +366,15 @@ final class OffsetPagination extends Widget implements PaginationWidgetInterface
     {
         $linkAttributes = $this->linkAttributes;
         if ($isDisabled) {
+            if (!array_key_exists('aria-disabled', $linkAttributes)) {
+                $linkAttributes['aria-disabled'] = 'true';
+            }
             Html::addCssClass($linkAttributes, $this->disabledLinkClass);
         }
         if ($isCurrent) {
+            if (!array_key_exists('aria-current', $linkAttributes)) {
+                $linkAttributes['aria-current'] = 'page';
+            }
             Html::addCssClass($linkAttributes, $this->currentLinkClass);
         }
         $link = Html::a($label, $url, $linkAttributes);